[Hcloud-Kernel] Share Memory
hcc_ipc_shm_destroy  구조체 변경
